chore: reduce memory occupation
by returning `dict` instead of `ParseResult` objects.
