+++ app/parser.py
@@ -1,11 +1,9 @@
 import asyncio
 import logging
 import os
+from dataclasses import asdict
 
-import numpy as np
 from ray import serve
-
-from .mineru_parser import MinerUParser, ParseResult
 
 logger = logging.getLogger(__name__)
 deployment_name = "DocumentParserDeployment"
@@ -20,6 +18,8 @@ deployment_name = "DocumentParserDeployment"
 class DocumentParser:
     def __init__(self):
         logging.basicConfig(level=logging.INFO)
+
+        from .mineru_parser import MinerUParser
 
         # Instantiate MinerUParser locally within this replica to do the work.
         # It doesn't need a parser_deployment_handle because it's the worker, not an orchestrator.
@@ -39,7 +39,7 @@ class DocumentParser:
         )
         logger.info("DocumentParser initialized.")
 
-    async def parse(self, data: bytes, filename: str) -> ParseResult:
+    async def parse(self, data: bytes, filename: str) -> dict:
         replica_context = serve.get_replica_context()
         logger.info(f"Replica {replica_context.replica_id} received parse request.")
 
@@ -52,7 +52,11 @@ class DocumentParser:
                 filename,
             )
             logger.info(f"Replica {replica_context.replica_id} completed parse.")
-            return result
+
+            # Do not return a ParseResult object directly. When Ray serializes this object,
+            # it will package the entire mineru_parser module,
+            # significantly increasing the memory overhead of the calling actor.
+            return asdict(result)
         except Exception as e:
             logger.error(
                 f"Replica {replica_context.replica_id}: Error during parse: {e}",
@@ -60,82 +64,82 @@ class DocumentParser:
             )
             raise
 
-    async def batch_image_analyze(
-        self,
-        images_with_extra_info: list[tuple[np.ndarray, bool, str]],
-        ocr: bool,
-        show_log: bool = False,
-        layout_model=None,
-        formula_enable=None,
-        table_enable=None,
-    ) -> list:
-        """
-        Remotely callable method to process a batch of images.
-        This method will be executed by a DocumentParser replica.
-        """
-        replica_context = serve.get_replica_context()
-        logger.info(
-            f"Replica {replica_context.replica_id} received batch_image_analyze "
-            f"request with {len(images_with_extra_info)} images."
-        )
+    # async def batch_image_analyze(
+    #     self,
+    #     images_with_extra_info: list[tuple[np.ndarray, bool, str]],
+    #     ocr: bool,
+    #     show_log: bool = False,
+    #     layout_model=None,
+    #     formula_enable=None,
+    #     table_enable=None,
+    # ) -> list:
+    #     """
+    #     Remotely callable method to process a batch of images.
+    #     This method will be executed by a DocumentParser replica.
+    #     """
+    #     replica_context = serve.get_replica_context()
+    #     logger.info(
+    #         f"Replica {replica_context.replica_id} received batch_image_analyze "
+    #         f"request with {len(images_with_extra_info)} images."
+    #     )
 
-        loop = asyncio.get_running_loop()
-        try:
-            # MinerUParser.batch_image_analyze calls may_batch_image_analyze, which is CPU/GPU intensive.
-            # Run it in an executor to avoid blocking the replica's asyncio event loop.
-            result_list = await loop.run_in_executor(
-                None,  # Use default ThreadPoolExecutor
-                self.mineru.batch_image_analyze,  # Call the method on the instance
-                images_with_extra_info,
-                ocr,
-                show_log,
-                layout_model,
-                formula_enable,
-                table_enable,
-            )
-            logger.info(
-                f"Replica {replica_context.replica_id} completed batch_image_analyze for {len(images_with_extra_info)} images."
-            )
-            return result_list
-        except Exception as e:
-            logger.error(
-                f"Replica {replica_context.replica_id}: Error during batch_image_analyze: {e}",
-                exc_info=True,
-            )
-            raise
+    #     loop = asyncio.get_running_loop()
+    #     try:
+    #         # MinerUParser.batch_image_analyze calls may_batch_image_analyze, which is CPU/GPU intensive.
+    #         # Run it in an executor to avoid blocking the replica's asyncio event loop.
+    #         result_list = await loop.run_in_executor(
+    #             None,  # Use default ThreadPoolExecutor
+    #             self.mineru.batch_image_analyze,  # Call the method on the instance
+    #             images_with_extra_info,
+    #             ocr,
+    #             show_log,
+    #             layout_model,
+    #             formula_enable,
+    #             table_enable,
+    #         )
+    #         logger.info(
+    #             f"Replica {replica_context.replica_id} completed batch_image_analyze for {len(images_with_extra_info)} images."
+    #         )
+    #         return result_list
+    #     except Exception as e:
+    #         logger.error(
+    #             f"Replica {replica_context.replica_id}: Error during batch_image_analyze: {e}",
+    #             exc_info=True,
+    #         )
+    #         raise
 
-    async def ocr(
-        self,
-        img: np.ndarray | list | str | bytes,
-        det: bool = True,
-        rec: bool = True,
-        mfd_res: list | None = None,
-        tqdm_enable: bool = False,
-    ) -> list:
-        """
-        Remotely callable method to perform OCR on a batch of images.
-        This method will be executed by a DocumentParser replica.
-        """
-        replica_context = serve.get_replica_context()
-        logger.info(f"Replica {replica_context.replica_id} received ocr request.")
+    # async def ocr(
+    #     self,
+    #     img: np.ndarray | list | str | bytes,
+    #     det: bool = True,
+    #     rec: bool = True,
+    #     mfd_res: list | None = None,
+    #     tqdm_enable: bool = False,
+    # ) -> list:
+    #     """
+    #     Remotely callable method to perform OCR on a batch of images.
+    #     This method will be executed by a DocumentParser replica.
+    #     """
+    #     replica_context = serve.get_replica_context()
+    #     logger.info(f"Replica {replica_context.replica_id} received ocr request.")
 
-        loop = asyncio.get_running_loop()
-        try:
-            # MinerUParser.ocr can handle a list of images.
-            ocr_results = await loop.run_in_executor(
-                None,  # Use default ThreadPoolExecutor
-                self.mineru.ocr,  # Call the method on the instance
-                img,
-                det,
-                rec,
-                mfd_res,
-                tqdm_enable,
-            )
-            logger.info(f"Replica {replica_context.replica_id} completed ocr.")
-            return ocr_results  # This should be a list of OCR results, one for each image in images_batch
-        except Exception as e:
-            logger.error(
-                f"Replica {replica_context.replica_id}: Error during ocr: {e}",
-                exc_info=True,
-            )
-            raise
+    #     loop = asyncio.get_running_loop()
+    #     try:
+    #         # MinerUParser.ocr can handle a list of images.
+    #         ocr_results = await loop.run_in_executor(
+    #             None,  # Use default ThreadPoolExecutor
+    #             self.mineru.ocr,  # Call the method on the instance
+    #             img,
+    #             det,
+    #             rec,
+    #             mfd_res,
+    #             tqdm_enable,
+    #         )
+    #         logger.info(f"Replica {replica_context.replica_id} completed ocr.")
+    #         return ocr_results  # This should be a list of OCR results, one for each image in images_batch
+    #     except Exception as e:
+    #         logger.error(
+    #             f"Replica {replica_context.replica_id}: Error during ocr: {e}",
+    #             exc_info=True,
+    #         )
+    #         raise
